+++ docs/changelog.rst
@@ -6,8 +6,11 @@ Changelog
 v0.5.0 (in development)
 -----------------------
 - Attachments' filenames can now be `None`
-- Added a `decompose()` function for decomposing
-  `~email.message.EmailMessage`\s into `MailItem`\s plus headers
+- Added a `decompose()` function for decomposing an
+  `~email.message.EmailMessage` into a `MailItem` plus headers
+- Added a `decompose_simple()` function for decomposing an
+  `~email.message.EmailMessage` into a text body, HTML body, attachments, and
+  headers
 - The ``subject`` argument to the `compose()` function & method can now be
   `None`/omitted
 - If an address argument to `compose()` is set to an empty list, the